Implementing Working SFX
Each Scene has a Sound Manager game object, that holds every sound effect that can play. In the menu screens, it's as simple as just having the buttons call the functions in the editor UI. With the actual gameplay scenes, the sounds must be played by calling the Manager script's appropriate function within the scripts of the other game objects. NOTE: In order to make it work, you need to also drag the manager object into the script's inspector when you define the manager script as a variable.